app complete

+++ frontend/src/components/WorkoutInfo.js
@@ -1,32 +1,49 @@
 import { useWorkoutsContest } from "../hooks/useWorkoutsContest";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 //date fns
-import formatDistanceToNow from "date-fns/formatDistanceToNow"
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
+const WorkoutDetails = ({ workout }) => {
+  const { dispatch } = useWorkoutsContest();
+  const { user } = useAuthContext();
 
-const WorkoutDetails = ({workout}) => {
-
-    const { dispatch } = useWorkoutsContest();
-
-    const handleClick = async () => {
-        const response = await fetch("/api/workouts/" + workout._id, {
-            method: "DELETE"
-        })
-        const json = await response.json();
-
-        if(response.ok){
-            dispatch({type: "DELETE_WORKOUT", payload: json})
-        }
+  const handleClick = async () => {
+    if (!user) {
+      return;
     }
+
+    const response = await fetch("/api/workouts/" + workout._id, {
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
+    });
+    const json = await response.json();
+
+    if (response.ok) {
+      dispatch({ type: "DELETE_WORKOUT", payload: json });
+    }
+  };
   return (
-    <div className='workout-details'>
+    <div className="workout-details">
       <h4>{workout.title}</h4>
-      <p><strong>Load (kg):</strong>{workout.load}</p>
-      <p><strong>Reps: </strong>{workout.reps}</p>
-      <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
-      <span className="material-symbols-outlined" onClick={handleClick}>Delete</span>
+      <p>
+        <strong>Load (kg):</strong>
+        {workout.load}
+      </p>
+      <p>
+        <strong>Reps: </strong>
+        {workout.reps}
+      </p>
+      <p>
+        {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
+      </p>
+      <span className="material-symbols-outlined" onClick={handleClick}>
+        Delete
+      </span>
     </div>
   );
-}
+};
 
 export default WorkoutDetails;
